[forever_new_ca] Add spider, separate forever_new_au_nz
